# server/docs/swagger.yaml
basePath: /
definitions:
...
  api.CompleteRequest:
    properties:
      attachments:
        items:
          $ref: '#/definitions/api.CreateMessageAttachmentRequest'
        type: array
      conversationId:
        type: string
      maxRetries:
        type: integer
      message:
        type: string
      providerOrder:
        items:
          type: string
        type: array
    type: object
  api.CompleteResponse:
    properties:
      attempts:
        type: integer
      output:
        type: string
      provider:
        type: string
      trace:
        $ref: '#/definitions/api.CompletionTraceResponse'
      usage:
        $ref: '#/definitions/api.TokenUsageDTO'
    type: object
  api.CompletionTraceResponse:
    properties:
      providerAttempts:
        items:
          $ref: '#/definitions/api.ProviderAttemptDTO'
        type: array
      runtimeActions:
        items:
          $ref: '#/definitions/api.RuntimeActionDTO'
        type: array
      toolCalls:
        items:
          $ref: '#/definitions/api.ToolCallDTO'
        type: array
    type: object
...
  api.CreateMessageAttachmentRequest:
    properties:
      createdAt:
        type: string
      dataUrl:
        type: string
      fileName:
        type: string
      id:
        type: string
      mimeType:
        type: string
      sizeBytes:
        type: integer
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.ProviderAttemptDTO:
    properties:
      attempt:
        type: integer
      error:
        type: string
      provider:
        type: string
      success:
        type: boolean
    type: object
...
  api.RuntimeActionDTO:
    properties:
      command:
        type: string
      kind:
        type: string
      output:
        type: string
      startedAt:
        type: string
      status:
        type: string
    type: object
...
  api.TokenUsageDTO:
    properties:
      completionTokens:
        type: integer
      promptTokens:
        type: integer
      totalTokens:
        type: integer
    type: object
  api.ToolCallDTO:
    properties:
      arguments:
        additionalProperties: {}
        type: object
      name:
        type: string
      output:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Self-hostable multi-device AI assistant API
  title: openCrow API
  version: "1.0"
paths:
...
  /v1/conversations/{id}/messages/{msgId}/regenerate:
    post:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: Message ID
        in: path
        name: msgId
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream of delta/tool_call/tool_result/done/error events
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Regenerate an assistant message via SSE stream
      tags:
      - conversations
...
  /v1/orchestrator/complete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Message and conversation ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CompleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CompleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Run a full orchestrator completion (blocking)
      tags:
      - orchestrator
  /v1/orchestrator/stream:
    post:
      consumes:
      - application/json
      parameters:
      - description: Message and conversation ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CompleteRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream of delta/tool_call/tool_result/done/error events
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Stream an orchestrator completion via SSE
      tags:
      - orchestrator
...
  /v1/realtime/last:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the last realtime event for the current user
      tags:
      - orchestrator
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"